// src/interfaces/IUser.ts
// Generated by https://quicktype.io

// Generated by https://quicktype.io

export interface IUserLogged {
    id:             number;
    fullname:       string;
    username:       string;
    email:          string;
    identification: string;
    phone:          string;
    state:          string;
    status:         boolean;
    created_at:     string;
    updated_at:     string;
    id_rol:         number;
    exp:            number;
}

// Generated by https://quicktype.io

export interface IUserByEmail {
    data: Data;
}

export interface Data {
    user: IUserSigned;
}

export interface IUserSigned {
    id:        string;
    firstName: string;
    lastName:  string;
    email:     string;
    is8base:   boolean;
    origin:    string;
    roles:     Roles;
    status:    string;
}

export interface Roles {
    items: Item[];
}

export interface Item {
    id:          string;
    name:        string;
    description: string;
}